Add files via upload
Task2:
Displays a bar chart of the Zweitstimmen result of the elections. Uses pyplot to draw the diagram.

The diagram looks similar to https://www.bundeswahlleiter.de/bundestagswahlen/2017/ergebnisse.html.
Uses color codes for the parties, and combines every party below 5% into Sonstige.
Displays the numeric values at the top of the bar.